Address review comments - -- support comma separated list of jsonfiles in test -- move v2 client stuff to sdk/client/v2, from sdk/v2client -- add x-binary-type headers to sendraw requests
