+++ examples/gen-addresses/main.go
@@ -6,10 +6,10 @@ import (
 	"fmt"
 
 	"github.com/algorand/go-algorand-sdk/client/kmd"
+	"github.com/algorand/go-algorand-sdk/client/v2/algod"
 	"github.com/algorand/go-algorand-sdk/crypto"
 	"github.com/algorand/go-algorand-sdk/future"
 	"github.com/algorand/go-algorand-sdk/types"
-	"github.com/algorand/go-algorand-sdk/v2client/algod"
 )
 
 // CHANGE ME
